Simplify challenge requirements
