+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "fe-home-assignment",
-  "version": "0.1.0",
+  "version": "1.0.0",
   "private": true,
   "description": "A frontend home assignment built with Next.js, TypeScript, and Tailwind CSS",
   "scripts": {
