+++ src/app/layout.tsx
@@ -13,8 +13,9 @@ const geistMono = Geist_Mono({
 });
 
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "FE Home Assignment",
+  description:
+    "A frontend home assignment built with Next.js, TypeScript, and Tailwind CSS",
 };
 
 export default function RootLayout({
