+++ src/components/HotelsSearchWidget/README.md
@@ -9,15 +9,14 @@ Your HotelsSearchWidget should include:
 1. **Place/Hotel Search**: An input field with autocomplete suggestions using the `/v1/suggest` endpoint
 2. **Date Selection**: Two date pickers for check-in and check-out dates
 3. **Guest Selection**: A numeric input for the number of guests
-4. **Search Action**: A search button that calls the `/v1/search` endpoint
+4. **Auto Search**: A search should be performed to the `/v1/search` endpoint automatically when the user changes any of the input fields.
 5. **Hotel List**: A list of hotels that match the search criteria and displays the hotel details received from the search request
 
 ### The Expected Outcome
 
 1. Implement your solution in this directory
-2. Add a new `widgets` page for this widget
-3. Include the widget as an iframe on the homepage
-4. Document your approach and any improvements made in the `NOTES.md` file
+2. Include the widget on the homepage in the provided section
+3. Document your approach and any improvements made in the `NOTES.md` file
 
 ---
